Collatz

Collatz

Top level Module

Index

Functions

Constants

Internal Functions

Internal Modules

Top level Module

Collatz.Collatz — Module

Provides the basic functionality to interact with the Collatz conjecture. The parameterisation uses the same (P,a,b) notation as Conway's generalisations. Besides the function and reverse function, there is also functionality to retrieve the hailstone sequence, the "stopping time"/"total stopping time", or tree-graph.

Examples

julia> import Pkg; Pkg.add("Collatz"); using Collatz

Index

Collatz.Collatz

Collatz._CC

Collatz._ErrMsg

Collatz._KNOWN_CYCLES

Collatz.__VERIFIED_MAXIMUM

Collatz.__VERIFIED_MINIMUM

Collatz.__assert_sane_parameterisation

Collatz.__initial_value_outside_verified_range

Collatz.__stopping_time_terminus

Collatz.collatz_function

Collatz.hailstone_sequence

Collatz.reverse_collatz_function

Collatz.stopping_time

Collatz.tree_graph