Internal Functions

Collatz.__assert_sane_parameterisation — Function

__assert_sane_parameterisation(P, a, b)

Handles the sanity check for the parameterisation (P,a,b) required by both the function and reverse function.

Args

P::Integer: Modulus used to devide n, iff n is equivalent to (0 mod P).

a::Integer: Factor by which to multiply n.

b::Integer: Value to add to the scaled value of n.

Collatz.__initial_value_outside_verified_range — Function

__initial_value_outside_verified_range(x)

Checks if the initial value is greater than __VERIFIEDMAXIMUM or less than __VERIFIEDMINIMUM. Only intended for the default parameterisation.

Args

x::Integer: The initial value to check if it is within range or not.

Collatz.__stopping_time_terminus — Function

__stopping_time_terminus(n, total_stop)

Provides the appropriate lambda to use to check if iterations on an initial value have reached either the stopping time, or total stopping time.

Args

n::Integer: The initial value to confirm against a stopping time check.

total_stop::Bool: If false, the lambda will confirm that iterations of n have reached the oriented stopping time to reach a value closer to 0. If true, the lambda will simply check equality to 1.